served/stayed/entered/exited/operating/beds_available_between re-work
An internal function `between_` has been added to perform the functions currently spanning the `*_between` family of functions. 
To retain the existing legibility of the between function taxonomy, the original function names are recycled as wrappers for `between_`.
With the exception of `operating` & `beds_available` the resulting output is the filtered `data.frame` such that these functions can be used in a pipe. 
For operating & beds_available the output is the logical vector resulting from the `if_else` statement - thus this can continue to be used within a `mutate` call to output a column.