Suggestion to make all ProjectType columns numeric in COHHIO_HMIS output
For consistency, it would make sense to have ProjectType be a numeric variable (it's native class) in all outputs from COHHIO_HMIS instead of a character in some places and a numeric in others (creating the likelihood for errors). Character coercion and comparison can occur on the app side after filtering since the character versions are only needed for human legibility. 

This branch is created to make this change.