Explicit use of environment when sourcing to eliminate reading RData
# Rationale
Current 00_daily_update `source`s scripts to the global environment by default, and clears it with `rm` after each `source` call. Inside of individual scripts, dependency objects are obtained by decompressing the saved RData files from previously sourced scripts, which is an expensive operation. We can omit the need for this by explicitly declaring environments for scripts to be sourced into and passing the necessary environment to the subsequently `source`d scripts that depend on it, thus eliminating the need to load the RData file. The code to load RData files is retained for working on scripts individually, though now it is run only when key objects from that script do not exist in the environment that the script is using.

These changes result in an approx. **20% reduction in the total run time for the script**.

# Changes
## 00_cohorts
 - condition for `load`
## 00_daily_update
### `stop_with_instructions`
 - function to `stop` the script and give users an informative error message
### increment
 - This is a time-keeping, progress reporting, and completion estimation function
 - at each `increment` it will:
   - print the number of `source` steps out of the total
   - give the elapsed time
   - if a previous run has been done by the user - it will give a % complete & an estimated time of completion
### Extraction
The `{archive}` package is used to automatically unzip the *7z* file so users don't have to worry about figuring out how to extract it themselves, and the archive is deleted automatically.

### `source` calls
 Each `source` call is now provided an environment in which to be executed that includes the objects the script depends on.

## 00_first_run
 - new dependencies added:
    - `{cli}`
    - `{rlang}`
    - `{archive}`
## 00_get_Export..
 - add condition for `load`
 - `source` must explicitly declare the environment inside the script or it sources to the global environment (and ends up unavailable to the script)
## 01-05, 08...
 - conditions for `load`
 - `source` to declared environment
## 07_SPMs
 - create the *random_data* directory if it isn't already created (avoids error)