Make data_prep self-contained, give accessor fn environment baseenv
accessor function were throwing an error to the effect of:
`function { not found`
This was because rlang::env(rlang::empty_env()) creates an environment which does not have the base package on the search path!
Remedied by using `rlang::env(baseenv())` ad the function environment.
Everything now encapsulated inside of data_prep now, so it's a self-contained function for writing data.frames to files and assigning accessor functions in their place.